market-vision-backend/requirements.txt
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
curl_cffi==0.10.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozendict==2.4.6
html5lib==1.1
idna==3.10
lxml==5.3.2
multitasking==0.0.11
numpy==1.26.4
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
protobuf==5.29.4
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
reportlab
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
websockets==15.0.1
yfinance==0.2.59
